DIAC-298 - framework for refactoring request fee remission to stop any concurrency from spilling into different pods (and also from different cases interacting with anothers data
