+++ src/main/java/uk/gov/hmcts/reform/iacaseapi/domain/service/RemissionDetailsAppender.java
@@ -13,8 +13,6 @@ import uk.gov.hmcts.reform.iacaseapi.domain.entities.ccd.field.IdValue;
 @Slf4j
 public class RemissionDetailsAppender {
 
-    private List<IdValue<RemissionDetails>> remissions;
-
     public List<IdValue<RemissionDetails>> appendAsylumSupportRemissionDetails(
         List<IdValue<RemissionDetails>> existingRemissionDetails,
         String feeRemissionType,
@@ -107,16 +105,5 @@ public class RemissionDetailsAppender {
         }
 
         return allRemissionDetails;
-    }
-
-    public void setRemissions(List<IdValue<RemissionDetails>> remissions) {
-        log.info("SETTING REMISSION: " + remissions);
-        log.info("TO: " + this.remissions);
-        this.remissions = remissions;
-    }
-
-    public List<IdValue<RemissionDetails>> getRemissions() {
-        log.info("GETTING REMISSIONS: " + remissions);
-        return remissions;
     }
 }
